Downloading the binaries using the python module instead of CURL
In some installations, CURL might not be updated, and might not have wget,
downloading from OSF will fail. Instead of using CURL or wget, we will use
the python request module with the security option to download the binaries.

The installation still depends on CURL or wget to download conda, if the
installation of conda fails, we have no other option then to recommend the
user to update their system.

Fixes #1107


Former-commit-id: 9e032f46fae70e3a1ba6105abc463c711dcdce28