Downloading the binaries using the python module instead of CURL
In some installations, CURL might not be updated, and might not have wget,
downloading from OSF will fail. Instead of using CURL or wget, we will use
the python request module with the security option to download the binaries.

The installation still depends on CURL or wget to download conda, if the
installation of conda fails, we have no other option then to recommend the
user to update their system.

Fixes #1107


Former-commit-id: 31a5969d30b06dc5c46c41c0fa7cfef5575d1ca0 [formerly 4f8950f701eb68eecf00875b300cc91cc12b0c27 [formerly 9e032f46fae70e3a1ba6105abc463c711dcdce28]]
Former-commit-id: 1acdb05495295d1c38027c35dad1ae46ef5a4c79
Former-commit-id: 954df227f88af1f0a50d2d5a8addb874aec50d67