sct_extract_metric: Do not zero negative values (#2049)
* sct_extract_metric: Do not zero negative values

Fixes #2048

* sct_extract_metric: Set label to 0 for outlier values

When extracting metrics, the previous implementation set data=0 for outlier data (inf, -inf, nan, etc.), which produced bias in the computed metric. Here, we set the label voxel to 0 instead so this voxel is not used during the computation of the metric.

* extract_metric: Added flag -discard-neg-val

This flag enables the user to discard or keep (default behavious) negative values.

* test_sct_extract_metric: Updated groud-truth

To accommodate the latest change in sct_extract_metric about handling of negative values

* batch_processing: Added flag -discard-neg-val

To introduce the concept to users.