aggregate_slicewise: Enabled to combine labels and fix MAP estimation
In the previous implementation, for MAP, the ML estimation was using the first label of the combined labels. For example, if user asked for metric in the spinal cord, the beta_0[0] (from ML) would be only based on the first label (WM), not on all the labels (in that case, WM + GM).