Normalize binaries' package name.
Previously we treated linux as two kinds of beasts: Debian and CentOS 6.
But we conflated Debian with all other linuxes, so really it was "linux"
and "centos6".

Going forward, I want to do builds the way PyPI does with manylinux [1]:
the oldest CentOS is considered linux full stop. With my recent rebuild,
that is CentOS 7. Update the package name to emphasize this clearly.

I picked CentOS 7 because that's the oldest version that ANTS will build
on, but I'm temporarily keeping a special case for CentOS 6 for the next
~6 months until it end-of-lifes.

[1] https://github.com/pypa/manylinux