Update motion correction snapshot tests.
In updating antsSliceRegularizedRegistration, and its samplingStrategy
to get the new version to be as deterministic as the 2016 version,
the estimates it gives on the test data have shifted. This updates the
snapshots we had of the values to the new values.

I am unsure if this is right. Maybe the old estimates were better?