Force single-threaded motion correction.
This is also to make the output deterministic, at the cost of running slow.

It turned out that using dense sampling wasn't enough; there was still some
numerical instability that came from the order of addition:

* https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues#variance-due-to-floating-point-precision-errors

For some reason it only appeared on OS X, and only about 10% of the time, and never on Linux.
I [showed](https://github.com/neuropoly/spinalcordtoolbox/pull/2642#issuecomment-623376962)
that the instability in isct_antsSliceRegularizedRegistration did exist on Linux, so something
still unknown about how we call it was hiding it there.

This was actually supposed to be in place already but the code had atrophied,
so all this does is fix it up.

See:

* https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues
* https://github.com/ANTsX/ANTs/issues/444#issuecomment-300927998
* https://github.com/ANTsX/ANTsR/issues/210#issuecomment-394171962