sct_check_dependencies: simplify validation logic.
This had a redundant inner loop inside it: get_dependencies() was called once
per each entry in get_dependencies(), at least, in certain cases.

I'm pretty sure the warning
  print_warning(more=(" (%s != %s reference version))" % (version, ver_pip_setup)))
was unreachable code.

I got rid of that and simplified it so now it always prints the
installed version number, if it can find it.