sct_run_batch: ensure sct_version gets logged to batch_log
This one was a bit tricky, I want to show the sct version
before parsing the arguments, in case there's a parse error,
but in order to setup the log, we need to parse the arguments.
So I duplicated the message, but for reasons that aren't immediately
obvious, I couldn't trap `int_sct` in the log (without reappearing
in the terminal output), so I had to manually print the version.