sct_run_batch.py: Adds email notifications
Adds 3 additional arguments to support email notifications
`email-to` the only required argument, `email-from` in case
the sender email is different than the recipient. And `-email-host`
with the smtp server:port to use.

Prompts for a password with getpass and offers to send a test email
to ensure the authentication succeeds. Emails on error and
success.