Various improvements: create log, send email, config file, include/exclude list of subjects (#2732)
* sct_run_batch.py: Convert to if name == main style

* sct_run_batch.py: Add init_sct() at the start of the script

* sct_utils.py: Add Tee class for splitting stderr and stdout

* sct_run_batch.py: Add extra inclusion/exclusion styles

Adds `(in/ex)clude-list` for accepting a space separated list
of subjects on the command line. and `(in/ex)clude-file` for
accepting a file.

* sct_run_batch.py: Add `-batch-log` for overall output log

* sct_run_batch: Move run single to toplevel for serialization

This is required for use with Pool.map

* sct_run_batch: ensure sct_version gets logged to batch_log

This one was a bit tricky, I want to show the sct version
before parsing the arguments, in case there's a parse error,
but in order to setup the log, we need to parse the arguments.
So I duplicated the message, but for reasons that aren't immediately
obvious, I couldn't trap `int_sct` in the log (without reappearing
in the terminal output), so I had to manually print the version.

* sct_run_batch: drop (in/ex)clude file in favour of config.json

The forthcoming config file will be the correct place to store
lists of subjects.

* sct_run_batch.py: Replace double with single quotes

* sct_utils: Remove __del__ from Tee

This method was interfering with pytest logging. I
think it is ok to omit because the fd objects will delete
themselves when the go out of scope. This was forcing the
closure of fds that may have been in use elsewhere

* sct_run_batch: Add json and yaml configuration files

This is a reasonably large commit with a few important things
- Adds a `-config` or `-c` argument
- This argument can read either json or yml configuration
- Semantics are commandline arguments override config, but a warning
  is issued that using this behaviour is discouraged (for hacking
  use only really)
- Modifies the argparse namespace to include the config modifications.
- Warns if there are unknown keys in the config
- Add tests for reading the config and some of the associated warnings/errors
- Add arguments to main, passing an array from sys.argv so that the
  main function can be used directly in the tests (avoids subprocess
  redirection/error-capture annoyances)

* sct_run_batch.py: Adds email notifications

Adds 3 additional arguments to support email notifications
`email-to` the only required argument, `email-from` in case
the sender email is different than the recipient. And `-email-host`
with the smtp server:port to use.

Prompts for a password with getpass and offers to send a test email
to ensure the authentication succeeds. Emails on error and
success.

* sct_run_batch.py: Absolutize `path_segmanual`

* sct_run_batch.py: Add config/argument logging

* sct_run_batch.py: Allow empty input to mean yes in send test email

As promised by the prompt.

* sct_run_batch: Add example configs and fix linewrap for R| strings

Adds example YAML and JSON config syntaxes for quick reference. Also
fixes incorrect handling of empty lines in R| strings when formatting
the help

* sct_run_batch: Switch task from a positional to named argument

* sct_run_batch.py: Fixed logical assertion

fixes https://github.com/neuropoly/spinalcordtoolbox/pull/2732#issuecomment-653600981

* Fixed passing of args inside sub-function

* sct_run_batch: change default to persist on error

Adds a boolean flag for configuring whether the pipeline should
abort at the first error or continue, with the default set to
continue.

* sct_run_batch: Fix tests for named task argument

* Display help if no argument is entered

* sct_run_batch: Fix missing return and typo

* sct_run_batch: Fix local variable lookup

* sct_run_batch: Fix no argument case

* sct_run_batch: Fix completed message

* sct_run_batch: fix return code check

* sct_run_batch: pep8 cleanup

* sct_run_batch: use explicit yaml loader to prevent warning

* sct_run_batch: extra newline on success message

* sct_run_batch: Mention possible yaml parse errors + recommend quotes

* Fixed cosmetic things in usage; changed default examples with jobs=-1

* Added info about gmail server config

* Modified how timing is displayed, added duration

* Fixed: checking for len(argv) instead of args

Because args is populated with default values, so the warning was always showed

Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>