Change setup.py for `pip install .` to fix Readthedocs CLI rendering (#2837)
* Treat scripts/ as package (for pip installation)

A hack is currently used involving `data_files` in setup.py. This causes
sct.compat.launcher to be unable to find the scripts when pip installed.
By treating scripts/ as a package, it will be installed to the right
location.

* setup.py: Use find_packages(), avoid data_files

This properly installs all of the subpackages of SCT, as well as the
CLI scripts (now treated as a package).