Use type factory function for parsing lists
nargs="+" could be used instead, but that would required delimiting
using a space, which would be a breaking change. By instead creating
more complex type parsing, the existing usage can be preserved.

This can be applied to many scripts, including ones that were previously
converted. Thus, this approach should be added to all relevant scripts
in a separate PR.