Add an -o (overwrite files) flag
----
EDIT 2022-05-17: This commit was originally called "Add a -c (clean) flag.".
But, the `-c` option is now in use by a different bit of functionality. So,
in rebasing this commit and fixing conflicts, I've renamed it to a unique
letter, and changed the language accordingly.     -@joshuacwnewton
----

This replaces -b (clean binaries) and -d (clean data), and expands
what they do to cover cleaning miniconda and the deepseg models.
With this, iterating on sct is a lot easier because the installer
doesn't take 3-10 minutes every time.

It should be safe to expand this to cover miniconda (in most case):
pip handles dependency versioning fine, and is smart enough to recognize
what's changed if anything has changed.

It is unsafe to use it with binaries/data/deepseg models because we do
not have those versioned. But it was already unsafe to use -b and -d in
anyway for exactly that reason; most of our users will be installing SCT to
fresh directories ~/sct_$VERSION each time anyway, so I think the tradeoff
is worth it.