Invert the -o flag (overwrite files) to be -k (keep files)
Instead of specifying a flag to turn *on* cleaning (i.e. overwrite the files),
we now specify a flag to turn *off* cleaning, which keeps the default as it was.

Also, move those prints to the relevant parts of the code instead of
just sitting at the cli; they're not that useful at the cli.
Now they only print if the particular piece is actually skipped so
install_sct describes what it's doing even with partial installs
(e.g. bin/ downloaded but conda deleted manually)