Factor out locking logic to @node_scoped.
This uses pickle to share. I don't know if that's a good idea.
This code got kind of crazy because wrapping fixtures is hard:
pytest does a bunch of introspection to magically come up with
data to feed into tests (or into other fixtures!) and I needed
to be compatible with that. Maybe there's a more reasonable way
involving, I dunno, making a separate fixture? Or something?

It also doesn't support generator-fixtures, which is unfortunate.
The mechanism to invoke them is tricky and not part of pytest's
public API. Maybe I'll give it a shot.