Added flag -s-template-id to use another segmentation (e.g. white matter) for registration to template (#3299)
* Added flag -s-template-id

Fixes #3297

* Fixed register_labdmark QC images not being output

A variable path_qc was required (but never passed) to be able to output useful QC plots for label-based registration. This commit fixes it by simply removing the path_qc variable everywhere in the file. These plots are used for advanced users anyway.

* Update spinalcordtoolbox/scripts/sct_register_to_template.py

* Update landmarks.py

Appease the linter

* Update sct_register_to_template.py

Appease the linter

* Clarified usage of s-template-id

Co-authored-by: Nick <nguenthe@uwaterloo.ca>