Drop `yes |` from the docs (#3307)
Use `-y` instead. `yes | ` is an old unix trick but it's also kind of fragile,
as I found when working on https://github.com/neuropoly/spinalcordtoolbox/pull/3125.

Also cull its last mention in our Dockerfile, not that we're maintaining or testing that.

And no need for `apt-get -y update`, it `update` doesn't prompt.