moco.py: Don't use the absolutepath, and instead copy files to tmpdir
The root of the problem is because 'moco.py' changes the cwd (making relative
paths unusable). Because the problem is in 'moco.py', it should be that
module's responsibility to handle the paths, NOT 'sct_dmri_moco'. In other words,
we shouldn't trust that incoming paths will always be absolute.

So, instead, copy the bvecs/bvals files to the tmpdir.

An alternate solution could be to not change the working directory at all, and to
simply prepend the tmpdir to the intermediate output files. This would avoid all
relative/absolute path issues. But, doing that would be a bigger refactor, so start
with the simpler solution for now. (Feedback welcome.)