Fix `-bvals` filepath handling and update bvals argparse help (#3376)
The root of the problem is because 'moco.py' changes the cwd to a temporary dir,
(making relative paths unusable). So, copy the bvecs and bvals files to the tmpdir, 
which avoids having to guess whether the path will be relative or absolute.

An alternate solution could be to not change the working directory at all, and to
simply prepend the tmpdir to the intermediate output files. This would avoid all
relative/absolute path issues. But, doing that would be a bigger refactor, so start
with the simpler solution for now.