zeros_like: Make a copy of the header
Because the exact header object was used, if the header in the original image
was changed, then the header of the `zeros_like` copy would be changed too.

This caused problems with sct_label_utils because reorienting the input image would mess up the header of the `zeros_like` overlay image.