Added unit test for separate sessions
This test is the basis of testing for the implementation of https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3413.
Currently, ses-XX folder within sub-XX folders will be ignored. The purpose of this test is to make sure they are considered as separate entities.
Note: currently the assert of the test is not implemented. It will be once the feature is implemented.