Make it possible to loop across "ses-" entity (#3414)
* Fixed PEP8

* Added unit test for separate sessions

This test is the basis of testing for the implementation of https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/3413.
Currently, ses-XX folder within sub-XX folders will be ignored. The purpose of this test is to make sure they are considered as separate entities.
Note: currently the assert of the test is not implemented. It will be once the feature is implemented.

* Fixed typo

* Implemented first stage of accounting for ses-XX sub-folder

* Parse subject/session in script runner

* Fixed bugs

* Working unit test for separate sessions

* Added test for BIDS with no session

* Removed code duplication via new function

Suggested in https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3414#discussion_r647517771

* Smarter creation/assertion of different sub/ses fields

Suggested by https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3414#discussion_r647564530