Refactor `register_step_ants_slice_regularized_registration`
This refactor splits the function into 2 subfunctions:
  * antsSliceRegularizedRegistration: A very light Python wrapper for the
    ANTs binary. This is meant to be a place where we can put logic that
    should be run on EVERY call to the binary.
  * slicereg_crop_extents: A bit of extra logic that was only used in
    registration. (It was skipped by `moco.py`, so it shouldn't be in the
    Python wrapper for the binary.)

The motivation for making a light wrapper is to stop having a run_proc(cmd)
call for every call to the binary. When we have individual calls to the binary,
it results in duplicating code snippets that need to run before/afterward. (For
example, the softmask fix.) If we have a wrapper, we can centralize this code.