Refactor `register_step_ants_registration`
This refactor splits the function into:
  * antsRegistration: A very light Python wrapper for the
    ANTs binary. This is meant to be a place where we can put logic that
    should be run on EVERY call to the binary.

The motivation for making a light wrapper is to stop having a run_proc(cmd)
call for every call to the binary. When we have individual calls to the binary,
it results in duplicating code snippets that need to run before/afterward. (For
example, the softmask fix.) If we have a wrapper, we can centralize this code.