Give up on trying to refactor softmask code between moco/register
I'm committing to preserve this work, just in case  I ever want to try
approaching this again in the future.

My plan was to try and apply the softmask in the same scope as when
we call the ANTs binaries (since the softmask is an ANTs-specific
limitation). That way, we could guarantee the softmask is applied in
a consistent way, regardless of when the binaries are called.

However, the way that softmasks were applied in `moco.py` is not consistent
with what I was planning on implementing for `register.py`, so unifying the
softmask code is much tricker (and likely not worth the effort to refactor).
So, in this case, I'm conceding that duplicating the softmask code is the
safest way to approach this issue.