Clean up `install/`, `issues/`, and `flirtsch/` folders in repo (#3461)
* `issues/`: Delete outdated user scripts
These scripts were developed 6 years ago to help a user on the
SourceForge help forum. They are out of date, and unlikely to be
of use to anyone presently.

* 'install/': Move inside `dev/` folder
Most of the scripts in this folder are outdated/unused/broken. This
used to be a module needed for SCT installation/updating, but is
no longer a part of the core SCT package, and is mostly a relic of
a previous era of SCT development. Therefore, it should be relegated
to the 'dev/' folder to avoid confusing new contributors.
NB: The 'dev/' folder itself is in need of cleanup, but that would be
a larger undertaking.

* Delete `flirtsch/` folder
As per @jcohenadad's suggestion, as the link in Tanguy's paper can't be changed, and the benefits outweigh the cons.