`aggregate_slicewise.py`: Let `csv.writer` handle str/float quoting
Previously, we converted *everything* to a string, which meant that
none of the fields would get quoted when written to CSV. This is what made it
 impossible for numpy to tell the non-numeric fields from the numeric fields.

So, instead of doing that, we leave the values in their original formats, then
specify `quoting=csv.QUOTE_NONNUMERIC` in `csv.writer`. This has the effect of
quoting all string values, but leaving the float values as they are.