Divide by the cosine, rather than multiply
The lesion is assumed to be parallel to the centerline.
The value `p_lst[2]` is the physical S-I thickness of the image
slice. To get the length parallel to the centerline, which should
be slightly larger, we have to divide by the cosine of the angle
between the centerline and the S-I direction, not multiply.