Replace `run_proc` with `main()` calls across multiple scripts
For some reason, I'm running into RuntimeErrors for these otherwise
benign calls. Perhaps it has something to do with subprocess?
But, this replacement is something we've been wanting to do anyway (for
other reasons), so this is a desirable change anyway.

For the record, the `register.py` call should probably be a call to `image.py`
if we can swing that, given the API/CLI slight desires.