`launcher.py`: Use `subprocess.call` instead of `os.execvpe`
On Windows, os.exec{vpe} doesn't function as one might expect on Linux. The call never returns,
so scripts hang indefinitely once they're finished executing. Further reading:

- Details about os.execvpe on Windows: https://stackoverflow.com/q/67371606
- Description of our specific use-case: https://stackoverflow.com/q/7004687

subprocess.call functions as we would expect it to. But, I haven't benchmarked or tested this
on Linux yet, so it's worth verifying that this does indeed do what we want it to.