`sys.py`: Refactor `check_exe` to match usage
In SCT's codebase, this function is only ever used to see if a program is accessible. The path string that gets returned by the function is never used once. So, this modifies the function to return boolean values instead.

Additionally, there is no need to manually check the PATH; shutil.which will do it for us. This also fixes a bug with Windows, because the previous custom method didn't take into account implicit file suffixes (e.g. `.exe`)