Fix a type error
Logically, `current_value` is a disc label, not a list index.
`list_disc_z_template` and `list_disc_value_template` are twin
lists, so using `list_disc_value_template.index()` is the way
to convert a disc label into a relevant list index.

This bug was not apparent because, at least for the default template,
`list_disc_value_template == [0, 1, 2, ..., 19]`.