`test_batch_processing.py`: Fix `skipif` to allow test to be run locally
Previously, the `skipif` locked the test to only be run on the CI. But, this is overly restrictive; you can still run the test locally, you just need make sure you've run `batch_processing.sh` first, because the output `.csv` files have to be generated.

So, we update the `skipif` to check for the presence of these `.csv` files.