`sct_download_data`: Use CI-packaged links for all 3 OSs
Note: In order to build ANTs on the `windows-2019` GitHub Actions CI runner,
it was necessary to bump the ANTs version to pull in an upstream ITK fix.

This meant that I also had to bump the ANTs version for both Ubuntu and OSX.
And, it's possible this has some unintended consequence on the registration results.
I want to test this using the test suite, but it's possible that we want to either

A) Keep the ANTs commits different for Windows and Ubuntu/macOS
B) Explore workarounds that don't require us to bump the ANTs version at all