Use CPU `--extra-index-url` to remove strict dependency pinning for `torch` (#3790)
To give our users a good experience we default
them to using torch's CPU builds, because otherwise
Linux users (which make up a a lot of our users?)
end up downloading the ~800MB GPU-compatible
torch builds, so we'd rather they get the smaller builds.

For two years now the best method we had for
doing this was `--find-links` + `torch==<version>+cpu`.
But this is very troublesome to maintain. The `+cpu`
syntax must include the rest of torch's version so this
duplicated (and when didn't pay close attention,
conflicted with) ivadomed's choices. torch isn't even
something we use directly, it's only for ivadomed.

But at some point recently torch put up full pip repos
(`--extra-index-url`) for each variant build:

* https://download.pytorch.org/whl/cpu
* https://download.pytorch.org/whl/cu101
* https://download.pytorch.org/whl/cu102
* https://download.pytorch.org/whl/cu111
* https://download.pytorch.org/whl/cu113
* https://download.pytorch.org/whl/lts/1.8/
* etc

This lets us unpin torch, because we can now direct
users to the CPU /build/ without having to
state which specific /version/ of torch to get.

This greatly simplifies maintenance and removes one
of the blockers towards https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/1526