`launcher.py`: Set PyTorch backend environment variables
NB: This is semi-redundant with the backend-setting lines inside of register.py. However, I couldn't think of a way off the top of my head to deduplicate this?

Since, if we *only* import in the launcher, then calling `sct_register_multimodal.main()` will bypass the launcher, resulting in an error.

But if we *only* import in `register.py`, we get the `sct_check_dependencies` error.