`create-release.yml`: Exclude `certifi @ file` from pip freeze
This is a temporary bugfix for an upstream conda issue. 

Right now, `pip freeze` is broken in conda environments, as it spits out nonsense file identifiers for certain packages. In our case, this affects only `certifi`, which is already installed by default in fresh conda environments, so it's safe to exclude from the `pip freeze`.

This fix was tested in https://github.com/spinalcordtoolbox/spinalcordtoolbox/pull/3798.