Remove obsolete 'noinspection' comments
These comments used to be necessary to silence a PyCharm code
inspection. The warning was caused by a test function argument being
named the same as a custom pytest fixture defined in the file (which
can't be helped: that's how pytest fixtures work). But from a quick
test on my laptop, it looks like PyCharm is now smart enough to
recognize this pattern, and doesn't warn about it anymore.