Fix executable file permissions and remove unnecessary header declarations (#3835)
* Make CLI scripts executable

* Make API module files not executable

* Remove shebang line from files that are not standalone scripts

These files only define classes and functions and global variables,
without actually *doing* anything, so they don't need a shebang line
(and don't have executable permissions anyway).

* Remove python2-style utf-8 declarations

They used to be necessary in python2, but utf-8 is the default in
python3. The only reason to keep them would be if someone's text
editor needed them; but since not all files contained them to begin
with, I doubt it.