`create-release.yml`: Ensure that `sys.platform` requirements are frozen
I'm not actually sure if this is the right fix here, since it's a pretty clumsy way of ensuring that the msvc-runtime requirements are frozen. Should we keep it unpinned?

That said, msvc-runtime isn't a Python module, it's a Microsoft Visual Studio DLL. And, from what I can tell, matplotlib isn't picky about the version: it only checks to see if the DLLs are present at all. So, my hunch is that trying to be particular about future updates is probably unnecessary. I think we can pin it and forget it.