Refactor release workflow into 3 stages to properly test PR branches  (#3836)
Before, the workflow was carried out in 1 job. This commit breaks that job into:

1. Generate requirements freeze
2. Testing
3. Release creation

The main goal here is to be able to isolate each of these steps so that we can uniquely choose the OS
and git branch we use for each step (since we want to release from master, but test from specific branches).

As a bonus, this also lets us more easily manage the actual release creation step, since we can use a single "if: github.event_name == 'workflow_dispatch'" conditional, instead of scattering the conditional across multiple steps.

With the new 3-job workflow, we can now make sure that we're NOT checking out master for the
first two jobs (generating requirements freeze, testing), and only checking out master
during the actual create-release job.

Before, this workflow was only run using `ubuntu-latest`. This commit expands the OSs
that we test to include macOS and Windows.