Pass `argv` to `generate_qc` functions (instead of `arguments` or `sys.argv[1:]`) (#3863)
* Replace `arguments` with `argv` in `generate_qc` calls

We want to use the unparsed arguments, because they show what the user actually passed,
rather than the Namespace object, which shows every possible argument.

* Replace `sys.argv[:1]` with `argv` in `generate_qc` calls

These are basically identical, but I figure it's best to not call `sys.argv[:1]` twice,
and instead use the actual value passed to `main()`. This standardizes the calls
across all of SCT's scripts.