Solution 3: `image.py`: Conditionally use `mmap=False` if OSErrors are thrown
This solution is a little more complex (and perhaps could be organized a bit better),
but the gist of it is to *not* unconditionally set `mmap=False`. Instead, the goal is to
only ever set `mmap=False` when we know that an OSError would be thrown.

This does a much better job at memory management. However, it means that we skirt
*right* up to the line of the acceptable # of open file descriptors. Meaning that, during processing, we would perhaps interfere with any other open processes that require open file descriptors.

I'm not entirely sure how much of a problem this is, though.