`sct_image`: Merge 2 redundant `fname_out` setting LOCs
We want `fname_out` to be properly set at the start of the script
(not the end of the script), so that we can access it during the
preparation of the QC report.

Moving `fname_out = fname_in[0]` means we also have to slightly
tweak the error thrown by `arguments.copy_header`.