`core.py`: Reorder conditional so that 'label' > 'rigid'
This is a simple fix for a bug stemming from the use of both
'type=label' and 'algo=rigid'. In such a case, both the 'if'
and 'elif' conditions are met. So the question is, should
the file have a '.txt' extension or a '.mat' extension?

On master, the conditional 'algo=rigid' takes priority over
'type=label', and the output file would have the '.mat' extension.

But, in the actual registration code, 'type=label' takes priority,
 and requires that the output file *should* have a '.txt' extension.
 So, we reorder the conditional such that 'type=label' takes priority
 over 'algo=rigid', properly assigning the expected '.txt' extension.

That said, I have qualms about how this code is organized in the
first place:

Right now, 'register()' determines which file extension to use
based on the parameters (algo, type, etc.). But, since we're
renaming the file, we *already have the expected file extension*!!

I presume we're not ever going to intentionally rename '.mat'
to '.txt' and vice versa. So, why not just preserve the extension,
rather than deriving the extension from the registration parameters?