`core.py`: Collapse '.mat' and '.txt' into single case
This is a rewritten solution that attempts to preserve the existing
file extension, rather than trying to infer the file extension from
the registration parameters.

I think this is a little more fool-proof, and explicitly prevents the
possibility from '.mat' being inadvertently renamed to '.txt' and vice
versa.

That said, this section of code still gives me pause:

- Why is it that the 'else:' case renames both warp_forward
  and warp_inverse, but the 'if:' case only renames 'warp_forward'?
- In that case, why even create the 'warp_inverse' variable at all?
- Does the 'warp_inverse' file just not exist?
- If we know that the 'warp_inverse' file doesn't get created in the
  '.mat' and '.txt' cases, then why not return e.g. 'None' for such
  a filename?

This requires a deeper dive into how the registration commands determine
their output files, though, and I don't want to balloon the scope of this
PR too muhc.