Pin `flake8<6.0.0` in our linting workflow (#3962)
* `lint_code.yml`: Pin `flake8<6.0.0`

I wonder if there's a better place to store this dependency. Could we have some sort of list of dev dependencies? Do devs rely on any other packages besides `flake8`? Local docs builds, maybe?

Maybe we could add an "extras" section to our setup.py for `flake8`, then install from there in both our CI workflow, and in our `flake8` git hook instructions in the Wiki.

* Escape version constraint

So that the `<` character is not interpreted by the shell as a redirection.

Co-authored-by: Mathieu Guay-Paquet <mathieu.guaypaquet@gmail.com>