`sct_deepseg`: Remove `deepseg.models.<function>` pattern
The following imports were problematic:

    import spinalcordtoolbox.deepseg as deepseg
    import spinalcordtoolbox.deepseg.models

Since the deepseg module only has 1 submodule (`models`),
these imports are redundant. We can access the submodule
directly using:

    from spinalcordtoolbox.deepseg import models

The reason I am making this commit first is because it caused
a linting error for the next commit (which is the true purpose
of this PR).