`launcher.py`: Remove `sct.__file__` pattern
This one is a little tricky! Right now, the `__file__`
property from `spinalcordtoolbox` actually points to the
top-level `__init__.py` file inside the package.

This is then used as a way to get the path of the
`spinalcordtoolbox` package's directory.

But, we don't need to import all of SCT just for this path!
So, I imported just `__file__`, and gave it a more descriptive
name.