`fs.py`: Make `basename` a mandatory argument
Since tmpdir names are logged as `info` messages, they're
at least somewhat intended to be read by users (due to the
default value of `-v` in all scripts).

But, using a bare tempdir name is unhelpful for users, as
it's unclear which process is actually generating the tempdir.

So, I think it would be better practice to make it mandatory to
use a basename, that way different temporary directories are
more easily distinguishable.

This also helps to expose scripts that don't specify `basename`.