`sct_propseg`: Fix incorrect `label-vertebrae` tmpdir name
This change contains 2 different aspects:
  1. `path_tmp` is only used in the context of the `use_optic`
     code-block, so it can be moved inside.
  2. The basename is flat-out wrong. It's not `label_vertebrae`,
     it's `centerline-optic` (in the context of `propseg`)

There's one extra note here:

I thought about re-using the *other* tmpdir that we create
earlier in the function (`folder_output_tmp`). The problem
with this, however, is that there's a LOC that copies all
files from the tmp directory to the output directory. This
is because `sct_propseg` will generate a variable number of
output files depending on the options passed to it, ao
rather than explicitly specifying which files should be
copied over, we just... lazily copy everything!

This means that the optic file would also be copied to
the output directory, which we don't want. So, re-using the
existing temporary directory would require a bit of a rewrite,
and I think this is a bit outside the scope of this PR.