`install_sct`: Isolate SCT conda by using `-p` instead of `-n`
Previously, when `-n` was used, there was a risk of the conda environment
being created in `$ENVS_DIRS/venv_sct`, rather than the expected
`$SCT_DIR/python/envs/venv_sct`.

This is because `-n` is supposed to be a *relative name* within the global conda
envs dir. This global envs dir can be set by users using a `.condarc` file. But,
SCT would prefer to ignore the global dir, and always create its env in $SCT_DIR.

Thankfully, `-p` specifies an absolute path, rather than a relative one, so it is
unaffected by the presence of `envs_dirs` in a `.condarc` file.