`sct_image.py`: Fix typo to address formatting issues
`fname_out` is a single filename string. `l_fname_out` is a list of component filenames.

The `display_viewer_syntax` function expects a list of filenames. When a single filename
string is passed, each char of the string is interpreted as a filename, hence the weird formatting.

Fixes https://github.com/spinalcordtoolbox/spinalcordtoolbox/issues/2526.