`metrics_to_PAM50.py`: Remove `levels_2_skip` variable
The purpose of this commit is to reduce the number of
lines of code, which saves readers some time/effort.

Rather than storing levels[0] and levels[-1] in a
variable, we can just access those values directly
from `levels` itself.

We also sort `levels`, that way we don't need to use
`min` or `max` -- we can just use [0] and [-1].