`metrics_to_PAM50.py`: Iterate through slices, not levels
The purpose of this commit is to pre-compute the slice
variables, which saves some lines of code inside the loop.

Also, later on, this will help us compute `scales` outside
the loop, which is important for letting us combine the
two loops.

(NB: The index `[::len(levels)-1]` looks complicated, but
don't worry, I'm going to remove that indexing later on
when I combine the two loops.)