`types.py`: BUG: Avoid accidentally mutating `points`/`derivatives`
Because the `compute_coordinate_system` used assignment (`=`) without
making a copy, the `/=` operation unexpectedly mutates `self.points` and
`self.derivatives`. I presume this was unintentional, and that the properties
should be preserved exactly as they were passed into the `Centerline` object.

We should be careful to verify that this is in fact a bug, as any downstream
usage of `self.points` and `self.derivatives` will change significantly.